crud
add

+++ mycontactbackend/controllers/contactController.js
@@ -51,6 +51,10 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -69,6 +73,10 @@ const deleteContact = asyncHandler(async (req, res) => {
   if(!contact){
     res.status(404);
     throw new Error("Contact not found");
+  }
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error("User don't have permission to delete other user contacts");
   }
   await Contact.deleteOne({_id: req.params.id});
   res.status(200).json({ contact });
